+++ sdk/ecalli/refine/invoke.ts
@@ -1,5 +1,5 @@
 /**
- * Ecalli 12: Invoke inner machine (Omega_V).
+ * Ecalli 12: Invoke inner machine.
  *
  * Run an inner PVM machine. The io_ptr points to a structure
  * containing gas limit and registers that is read before and
@@ -18,4 +18,4 @@
  */
 // @ts-expect-error: decorator
 @external("ecalli", "invoke")
-export declare function invoke(machine_id: u32, io_ptr: u32, out_r8: usize): i64;
+export declare function invoke(machine_id: u32, io_ptr: u32, out_r8: u32): i64;
